Add the Command.Name field
Add the Name field to the Command type, to use directly instead of
deducing the command name from the UsageLine field.
Add the parent field to keep track of the command ancestors.

Remove the Name global variable and update the Run function.

Remove the TestName test suite.

Improve the name used in the sub test.