Add the Parse function
The Run function does not allow the caller to control error handling
(Run will call os.Exit in case of errors) and the output written to
stderr.

Add the Parse function, that accept a Command, a FlagSet and a
command-lint argument list, and return the requested command and it's
arguments.

Change the Run function to call Parse with the Main command and
flag.CommandLine and handle the errors.