Improve the configure function
Change the configure function to accept a *Command instead of a
*flag.FlagSet.

Call FlagSet.Init with Command.String, in order to use the full command
name.

In the restore function, set FlagSet.Usage to Command.Usage.

The latter two changes are not strictly necessary, but only in order to
make Command.Flag consistent.