Rename Command.UsageFunc to Usage
Rename Command.Usage to Command.usage.

Rename Command.UsageFunc to Command.Usage, and update the documentation.
The Command.Usage field is more convenient and coherent for the user.

Update the documentation of Command.Long field.

Update the configure function and its documentation.

Update the Run function to use Command.usage.